+++ backend/app/adapter/recaptcha/recaptcha.go
@@ -5,12 +5,12 @@ import (
 	"net/http"
 
 	"github.com/short-d/app/fw/webreq"
-	"github.com/short-d/short/backend/app/usecase/external"
+	"github.com/short-d/short/backend/app/usecase/requester"
 )
 
 const verifyAPI = "https://www.google.com/recaptcha/api/siteverify"
 
-var _ external.ReCaptcha = (*Service)(nil)
+var _ requester.ReCaptcha = (*Service)(nil)
 
 // Service consumes with Google ReCaptcha V3 APIs through network.
 // https://developers.google.com/recaptcha/docs/verify
@@ -20,16 +20,16 @@ type Service struct {
 }
 
 // Verify checks whether a captcha response is valid.
-func (r Service) Verify(captchaResponse string) (external.VerifyResponse, error) {
+func (r Service) Verify(captchaResponse string) (requester.VerifyResponse, error) {
 	headers := map[string]string{
 		"Content-Type": "application/x-www-form-urlencoded",
 	}
 	// TODO(issue#739): Change ReCaptcha API body to a map.
 	body := fmt.Sprintf("secret=%s&response=%s", r.secret, captchaResponse)
-	apiRes := external.VerifyResponse{}
+	apiRes := requester.VerifyResponse{}
 	err := r.http.JSON(http.MethodPost, verifyAPI, headers, body, &apiRes)
 	if err != nil {
-		return external.VerifyResponse{}, err
+		return requester.VerifyResponse{}, err
 	}
 	return apiRes, nil
 }
