+++ backend/app/adapter/recaptcha/recaptcha_integration_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/short-d/app/fw/assert"
 	"github.com/short-d/app/fw/webreq"
-	"github.com/short-d/short/backend/app/usecase/external"
+	"github.com/short-d/short/backend/app/usecase/requester"
 )
 
 func TestReCaptcha_Verify(t *testing.T) {
@@ -23,7 +23,7 @@ func TestReCaptcha_Verify(t *testing.T) {
 		name         string
 		httpResponse *http.Response
 		httpErr      error
-		expRes       external.VerifyResponse
+		expRes       requester.VerifyResponse
 	}{
 		{
 			name: "successful request with score = 0.8",
@@ -39,7 +39,7 @@ func TestReCaptcha_Verify(t *testing.T) {
 }
 `,
 				)))},
-			expRes: external.VerifyResponse{
+			expRes: requester.VerifyResponse{
 				Success:       true,
 				Action:        "homepage",
 				Score:         0.8,
