+++ backend/app/usecase/requester/verifier.go
@@ -1,10 +1,8 @@
 package requester
 
-import "github.com/short-d/short/backend/app/usecase/external"
-
 // Verifier verifies in coming network to prevents cyber attacks.
 type Verifier struct {
-	service external.ReCaptcha
+	service ReCaptcha
 }
 
 // IsHuman checks whether the request is sent by a human user.
@@ -17,7 +15,7 @@ func (r Verifier) IsHuman(recaptchaResponse string) (bool, error) {
 }
 
 // NewVerifier creates new request verifier.
-func NewVerifier(service external.ReCaptcha) Verifier {
+func NewVerifier(service ReCaptcha) Verifier {
 	return Verifier{
 		service: service,
 	}
