+++ backend/go.mod
@@ -3,7 +3,6 @@ module github.com/short-d/short/backend
 go 1.13
 
 require (
-	github.com/DATA-DOG/go-sqlmock v1.4.1 // indirect
 	github.com/golang/protobuf v1.4.0 // indirect
 	github.com/google/wire v0.4.0
 	github.com/graph-gophers/graphql-go v0.0.0-20200309224638-dae41bde9ef9 // indirect
